Extract depthwise conv into a sub-module
